【AC】[AtCoder Grand Contest 047]A - Integer Product

+++ agc047/agc047_a.cpp
@@ -95,7 +95,7 @@ int main() {
     rep(i, n) {
         pairs.push_back(P(tows[i], fives[i]));
     }
-    map<P, int> counts;
+    map<P, ll> counts;
     for (P p : pairs) counts[p]++;
     ll ans = 0;
